add test for readCounters()
Improve the error handling for cullCounters() and readCounters().  Move the
value of BUFSIZE to the header for the tests and give it a more meaningful name.

Writing to a file descriptor and then reading from it without first closing the
file necessitates first flushing the write and then seeking back to the
beginning of the file.