change data structures to map and set
std::vector is an inefficient container when code needs to erase elements, as
the removal of an element in the middle of a container triggers a copy of all
subsequent elements.  Also, std:set_difference() requires a sorted set, so the
function is likely creating a temporary sorted copy of the vector each time it
is invoked.  Therefore change the data structure which holds struct pcounter
elements to a std::map, choosing the PID associated with each counter as the map
key.  At the same time, change from a range loop to an interator for the
cullCounters() function, which need not then invoke std::find().

Also change the container which holds PIDs to a std::set, with the benefit that
no element can be included twice.